requirements.txt
django==2.0.7
django-select2>=6.0.2
psycopg2>=2.7.4
junos-eznc>=2.1.7
pyvmomi>=6.7.0
django-ajax-selects>=1.7.0
django-appconf>=1.0.2
django-smart-selects>=1.5.3
jinja2>=2.9.6
pycparser>=2.17
pyOpenSSL>=17.3.0
pyparsing>=2.2.0
requests>=2.18.4
urllib3>=1.22
djangorestframework>=3.8.2
djangorestframework-jwt>=1.11.0
psycopg2-binary
django-background-tasks